note: support republishing a note to all connected relays
while testing duplicated relies the note with the funny replies
was only available on damus which seemed to be under heavy load.
added a republish button to be able to move the note to our relay
for testing purpose.

republishing should probably be more finetuned to be able to
publish to individual relays, but that requires relay management
first. hiding this feature for the moment.